+++ src/main/java/io/cavia/trader/module/auth/dto/LoginRequestDto.java
@@ -1,0 +1,14 @@
+package io.cavia.trader.module.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+
+@Getter
+public class LoginRequestDto {
+
+    @NotBlank(message = "이메일은 필수 입력 값입니다.")
+    private String username;
+
+    @NotBlank(message = "비밀번호는 필수 입력 값입니다.")
+    private String password;
+}
