--- src/main/java/io/cavia/trader/module/member/controller/PasswordViewController.java
@@ -1,14 +1,0 @@
-package io.cavia.trader.module.member.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class PasswordViewController {
-
-    @GetMapping("/forgot-password")
-    public String showPasswordForm() {
-        return "member/reset-password";
-    }
-
-}
