+++ src/main/java/io/cavia/trader/module/member/dto/NicknameUpdateRequestDto.java
@@ -1,0 +1,12 @@
+package io.cavia.trader.module.member.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class NicknameUpdateRequestDto {
+    private String nickname;
+}
