+++ src/main/java/io/cavia/trader/module/member/dto/PasswordRequestDto.java
@@ -1,0 +1,12 @@
+package io.cavia.trader.module.member.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
+public class PasswordRequestDto {
+    private String password;
+}
