--- src/main/java/io/cavia/trader/module/member/mypage/dto/NicknameUpdateRequestDto.java
@@ -1,12 +1,0 @@
-package io.cavia.trader.module.member.mypage.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-@AllArgsConstructor
-public class NicknameUpdateRequestDto {
-    private String nickname;
-}
