--- src/main/java/io/cavia/trader/module/member/mypage/service/MyPageService.java
@@ -1,24 +1,0 @@
-package io.cavia.trader.module.member.mypage.service;
-
-import io.cavia.trader.module.member.entity.Member;
-import io.cavia.trader.module.member.mypage.dto.GameParticipationDto;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface MyPageService {
-
-    public List<GameParticipationDto> findByMemberId(int memberId);
-
-    public Member findById(Long id);
-
-    public void changeNickname(Long id, String nickname);
-
-    public boolean validatePassword(Long id, String password);
-
-    public int changePassword(Long id, String password, LocalDateTime passwordUpdatedAt);
-
-    public int resetCash(Long id);
-
-    public int deleteMember(Long id, String password);
-}
