+++ src/main/resources/mapper/member/EmailVerificationMapper.xml
@@ -3,9 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="io.cavia.trader.module.member.repository.EmailVerificationMapper">
+<mapper namespace="io.cavia.trader.module.auth.repository.EmailVerificationMapper">
 
-    <insert id="saveOrUpdate" parameterType="io.cavia.trader.module.member.entity.EmailVerification">
+    <insert id="saveOrUpdate" parameterType="io.cavia.trader.module.auth.entity.EmailVerification">
         INSERT INTO email_verifications (
         email,
         verification_key,
@@ -24,7 +24,7 @@
         created_at = VALUES(created_at)
     </insert>
 
-    <select id="findByEmail" parameterType="string" resultType="io.cavia.trader.module.member.entity.EmailVerification">
+    <select id="findByEmail" parameterType="string" resultType="io.cavia.trader.module.auth.entity.EmailVerification">
         SELECT
         email,
         verification_key,
