+++ src/main/resources/mapper/mypage/MypageMapper.xml
@@ -2,8 +2,8 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="io.cavia.trader.module.member.mypage.mapper.MyPageMapper">
-    <select id="findByMemberId" resultType="io.cavia.trader.module.member.mypage.dto.GameParticipationDto">
+<mapper namespace="io.cavia.trader.module.member.mapper.MyPageMapper">
+    <select id="findByMemberId" resultType="io.cavia.trader.module.member.dto.GameParticipationDto">
         select
         id,
         member_id as memberId,
